+++ Assets/OxGKit/InfiniteScrollView/Scripts/Runtime/Core/ScrollView/HorizontalInfiniteScrollView.cs
@@ -71,7 +71,7 @@ namespace OxGKit.InfiniteScrollView
                             if (this._cellList[i] == null)
                             {
                                 if (this._cellPool.Count > 0) cell = this._cellPool.Dequeue();
-                                else Logging.Print<Logger>("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
+                                else Logging.PrintError<Logger>("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
                             }
                             // Check cell direction pivot
                             float dirCoeff = 1f;
@@ -99,7 +99,7 @@ namespace OxGKit.InfiniteScrollView
                             if (this._cellList[i] == null)
                             {
                                 if (this._cellPool.Count > 0) cell = this._cellPool.Dequeue();
-                                else Logging.Print<Logger>("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
+                                else Logging.PrintError<Logger>("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
                             }
                             // Check cell direction pivot
                             float dirCoeff = 1f;