+++ Assets/OxGKit/SaverSystem/Scripts/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG
 
+## [1.0.1] - 2025-07-10
+- Added Dispose method to Saver.
+
 ## [1.0.0] - 2025-04-10
 - Added SaverSystem.