+++ Assets/OxGKit/TimeSystem/Scripts/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+## [1.1.1] - 2025-07-10
+- Organized print level.
+
 ## [1.1.0] - 2025-05-15
 - Added NtpTime support for HTTP protocol on WebGL (using timeapi.io)